Add ability to calculate hours working past midnight
