+++ main.py
@@ -1,29 +1,4 @@
-def convertInputToHours(userInp):
-	allHoursWorked = []
-	for hour in userInp:
-		workedHours = hour.split(" -> ")
-		assert(len(workedHours) == 2)
-
-		leftTime = workedHours[0].split(":")
-		rightTime = workedHours[1].split(":")
-
-		leftHour = int(leftTime[0])
-		leftMin = int(leftTime[1])
-
-		rightHour = int(rightTime[0])
-		rightMin = int(rightTime[1])
-
-		minutesWorked = rightMin - leftMin
-		if minutesWorked < 0:
-			minutesWorked = 60 + minutesWorked
-			leftHour += 1
-
-		hoursWorked = rightHour - leftHour
-
-		timeWorked = str(hoursWorked) + "h" + str(minutesWorked) + "m"
-		allHoursWorked.append(timeWorked)
-
-	return allHoursWorked
+import inputToHours
 
 print("Enter hours worked throughout each shift in the format 5:12 -> 9:32:\n");
 
@@ -38,6 +13,6 @@ while (userInput != ""):
 	rawHours.append(userInput)
 	userInput = input("Shift " + str(activeShift) + ": ")
 
-hours = convertInputToHours(rawHours)
+hours = inputToHours.convertInputToHours(rawHours)
 
 print(hours)
